requirements.txt
streamlit
azure-ai-vision-imageanalysis
openai
python-dotenv
pandas